add initial support for crossing suar boundries
    python3 ./qmain 112:1
recite from the 1st aya of al-Ikhlaas (Sura 112), to the end of
the Quran.

    python3 ./qmain 112:1 113:2
recite from the 1st aya of al-Ikhlaas (Sura 112), to the 2nd aya of
al-Falaq (Sura 113), inclusive. A Basmala between them, on a line by
itself, is needed.

    python3 ./qmain 78:1
recite the entire Juz of Amm, the 30th part of the Quran.

    python3 ./qmain 67:1 77:-1
recite the entire Juz of Tabaarak, the 29th part of the Quran.

    python3 ./qmain 2:1 2:-1
is the same as
    python3 ./qmain 2
and as
    python3 ./qmain 2 1
which makes you recit the entire 2nd sura.

This commit is additive and backward-compatible. The 1st arg must
contain a colon (':'). So must the 2nd arg. The 2nd arg is optional
and defaults to the end of the Quran.

In this mode, an aya can be negative, to count from the end:

    python3 ./qmain 2:1 2:-1   # the entire sura
    python3 ./qmain 2:1 2:-2   # the entire sura but the last aya
    python3 ./qmain 2:1 2:-10  # the entire sura but the last ten ayat

Note that an aya of zero is an error. And also note that only this mode
supports the negative ayat; the normal single-sura mode doesn't (yet).

A Basmala is added between the suar. Not before the 1st one, even when
starting from the first aya. So `2:1 3:-1` would have only one Basmala.

The Basmala, currently, interacts badly with pagebreaks, because
pagebreaks, are associated by the next aya, not the previous, thus
you will type a sura, the Basmala, one or two blank lines, then the next
sura. That will be fixed later.