switch a tajweed-color-coded text
I made a side project to colorize the Quran text according to the
Tajweed rules, a la Mushaf Dar-ul-Marefa (easyquran.com).
That project will be completely published soon, but for now, only
the current resulting colorized text is used in Recite Web.

Notes:
- The entire text is pre-colorized, instead of repeatedly colorizing
  it in JavaScript.
- The color tags are compact and language independent; they are changed
  to HTML/CSS just before using.
- The color tags are changed into "<span_class=...>", and after
  splitting the ayat into words, they are changed to "<span class=...>".

I also fixed going backward, because the introduction of the tags broke
my previous code.

Also, I added two options in the interface: one for selecting the text
color, and one for the end of ayah sign and the number. I updated the
help accordingly. And fixed the keypress handling.