enter: support movement by "chunks"
("enter" means "Recite Enter", an internal name for the non-"txt" proj)

A "chunk" is a part of an aaya, which you can pause after.

An aaya may be one chunk (an aaya with no waqf signs),
two chunks (an aaya with exactly one waqf, that is not at the end of
the aaya), three chunks, or more.

A chunk does not cross the aaya boundary.

The waqf signs are Sakta, Wasl Awla, Waqf Awla, Waqf Lazem, Waqf Jaez,
and Split Waqf. See kind_of_portion() in javascript.js for details.

An aaya is split on the waqf signs to get its chunks.

I use the work "chunk" for lack of a better word; I'll hopefully change
it some time soon.