make Escape repeat the current ayah audio
It's primarily a workaround for a WONTFIX issue:
using the url params to select ayat, a qari, and
teacher mode, makes the webpage play audio before
any user interaction, which all modern browser
forbid by default. So we need a quick and easy
way to re-play the current ayah audio recitation.