add printed-like lines
Now the lines can be laid out like the printed moṣħafs.

A workaround is needed for Blink and is introduced.

An experimental gap-after-waqf-sign feature is also introduced.
Currently it's a very crude implementation and far from being
identical to the behavior of Moṣħaf Dar-ul-Ma’rifa, but close
enough for many uses.

The screenshot and OpenGraph and Twitter metadata are updated.