Auto-emulate the mainstream Arabic KB
If a user starts typing with a non-Arabic keyboard layout,
Recite now (whether in an Imlaai quiz or in a search bar)
will automatically emulate the mainstream Arabic layout (IBM PC)
for them.

If any emulation is selected by the user (through the URL params;
see the README), then this auto-emulation is never attempted.